% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detach_Packages.R
\name{detachPackages}
\alias{detachPackages}
\title{Detach Packages}
\usage{
detachPackages(package.list)
}
\description{
This function detaches all but the main (basic) packages 
       pre-loaded in NIDAP Code Workbook.
}
\details{
This method will detach packages that are not considered basic 
       packages required by Palantir's Code Workbook, and reattach new 
       packages needed by the current template
}
