% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_File_Path.R
\name{getFilePath}
\alias{getFilePath}
\title{Get file paths}
\usage{
getFilePath(data, ext, type = "r")
}
\arguments{
\item{ext}{Extension of file, for instance, "xlsx", "csv", "'txt, "RDS".}

\item{type}{Data type, by default, 'r'.}

\item{dataset}{Data usually in transform input format to import}
}
\value{
File path in local node to read into code template/transform
}
\description{
This function gets specific file paths for loading files into 
       NIDAP
}
\details{
This method will copy files to compute node for reading and returns
       the local file path for reading into code workbook.
}
