% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loop_Unzip.R
\name{loopUnzip}
\alias{loopUnzip}
\title{Unzip files}
\usage{
loopUnzip(dir)
}
\arguments{
\item{dir}{File path for zipped data file}
}
\value{
Local file path for unzipped data file
}
\description{
This function unzips specific file paths for loading files into 
       NIDAP
}
\details{
This method will unzip files for reading and returns
       the unzipped file path for reading into code workbook.
}
