% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-files.R
\name{nidapGetFiles}
\alias{nidapGetFiles}
\title{Get Files}
\usage{
nidapGetFiles(data, ext, type = "r")
}
\arguments{
\item{ext}{File extension. For instance, "xlsx", "csv", "'txt, "rds"}

\item{type}{Data type, by default, 'r'}

\item{dataset}{Dataset to import, usually in transform input format}
}
\value{
File name(s) to read into code template or code transform
}
\description{
This function returns specific file paths for an input dataset  
      in transform input format containing either a single or multiple files.
}
\details{
This method will copy files to the compute node and returns
       the original file names for reading into code workbook.
}
