% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Path.R
\name{nidapGetPath}
\alias{nidapGetPath}
\title{Load Data}
\usage{
nidapGetPath(input, name, type = "r")
}
\arguments{
\item{input}{transorm input data to import}

\item{name}{name of rds file}

\item{type}{data type, by default, 'r'.}
}
\value{
data read from the rds file
}
\description{
This function loads specifically an RDS file as the input data 
       format.
}
\details{
This method will import RDS data into a template/code transform
       in Code Workbook.
}
