% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Packages.R
\name{nidapLoadPackages}
\alias{nidapLoadPackages}
\title{Load Packages in NIDAP}
\usage{
nidapLoadPackages(current.packages, suppress = TRUE)
}
\arguments{
\item{suppress}{boolean to suppress messages.  By default set to TRUE}

\item{current.package}{vector of packages that are being used in the template}
}
\description{
This function will load packages and reorder package priorities
       while suppressing startup messages by default.
}
\details{
This method provides a way to automatically (optionally) suppress 
       package startup messages in NIDAP and removes all other unnecessary 
       packages
}
