% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reorder_Packages.R
\name{reorderPackages}
\alias{reorderPackages}
\title{Reorder Packages}
\usage{
reorderPackages(package.list)
}
\arguments{
\item{package.list}{packages currently being used in code template}
}
\description{
This function detaches all but the main (basic) packages 
       pre-loaded in NIDAP Code Workbook and reattaches the current packages
       to prioritize them for usage in the active code template.
}
\details{
This method will detach packages that are not considered basic 
       packages required by Palantir's Code Workbook, and reattach new 
       packages needed by the current template
}
