% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_RDS.R
\name{writeRDS}
\alias{writeRDS}
\title{Write RDS file}
\usage{
writeRDS(data, filename = "file.RDS")
}
\arguments{
\item{data}{data to be saved into an RDS file}

\item{filename}{filename for RDS file, "out.RDS" by default.}
}
\description{
This function saves an RDS file as output data
}
\details{
This method will save data into an RDS file in Code Workbook.
}
