% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop-unzip.R
\name{loopUnzip}
\alias{loopUnzip}
\title{Unzip Files in a Loop}
\usage{
loopUnzip(dir)
}
\arguments{
\item{dir}{File path for zipped data file(s)}
}
\value{
Local file path for unzipped data file(s)
}
\description{
This function unzips specific file(s) in a file path
}
\details{
This method will return unzipped file(s) for reading into a code 
     workbook template/transform.
}
