% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-files.R
\name{nidapGetPath}
\alias{nidapGetPath}
\title{Get Path}
\usage{
nidapGetPath(input, name, type = "r")
}
\arguments{
\item{input}{transform input dataset to import}

\item{name}{name of input file, usually transform input format}

\item{type}{data type, by default, 'r'.}
}
\value{
file path from which to read in data. Note that the original file 
   name will not be retained
}
\description{
This function returns the file path from a transform input 
       dataset in NIDAP Code Workbook
}
\details{
This function will return a file path from a transform input dataset 
       into a code template or code transform in the NIDAP instance of
       Foundry Code Workbook.
}
